% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KGSS-package.R
\docType{package}
\name{KGSS-package}
\alias{KGSS}
\alias{KGSS-package}
\title{KGSS: Data Package of Korean General Social Survey}
\description{
Korean General Social Survey Data from http://kgss.skku.edu/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/KGSS}
  \item Report bugs at \url{https://github.com/mrchypark/KGSS/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
